requirements.txt
anyio==3.5.0
ariadne==0.14.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
graphql-core==3.1.7
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
